backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
Pillow>=10.0,<12.0
google-auth>=2.36,<3.0
requests>=2.32,<3.0
